Improve operator button behaviors
When a second operator is pressed after a complete operation has been supplied
(2 + 2 + ) the previous operation is calculated and displayed in the result
div. If enter is pressed, the result can be used immediately in a following operation.

Currently pressing a number button after hitting enter does not replace the contents of
the display window/start a new operation. That needs to be corrected.